2.3 front ref Выделен store для таблицы.
