+++ frontend/src/components/PatientHistory/PatientCard/PatientCard.style.css
@@ -47,7 +47,6 @@
     gap: 4px;
 }
 
-.btn-edit-treatment,
 .btn-save-treatment,
 .btn-cancel-treatment {
     padding: 6px 16px;
@@ -56,14 +55,6 @@
     cursor: pointer;
     font-size: 14px;
     font-weight: 500;
-}
-
-.btn-edit-treatment {
-    background: #3498db;
-    color: white;
-}
-.btn-edit-treatment:hover {
-    background: #2980b9;
 }
 
 .btn-save-treatment {
@@ -102,4 +93,23 @@
     border-left: 3px solid #e74c3c;
     border-radius: 4px;
     white-space: pre-line;
+}
+
+.treatment-section-header .btn-edit-treatment {
+    background: transparent;
+    border: 1px solid #3498db;
+    color: #3498db;
+    font-size: 18px;
+    width: 32px;
+    height: 32px;
+    border-radius: 4px;
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 0;
+}
+.treatment-section-header .btn-edit-treatment:hover {
+    background: #3498db;
+    color: white;
 }
