+++ frontend/src/components/PatientHistory/PatientHistory.script.js
@@ -1,6 +1,4 @@
 import { extractIdFromIri } from '@/utils/apiHelpers';
-import apiClient from '@/api/client';
-import { testHistoryApi } from '@/api/testHistory';
 import { useAppointmentAddStore } from '@/stores/appointmentAddStore';
 import AppointmentAdd from '@/components/PatientHistory/AppointmentAdd/AppointmentAdd.vue';
 import TestAddModal from '@/components/PatientHistory/TestAddModal/TestAddModal.vue';
@@ -9,6 +7,7 @@ import PatientCard from '@/components/PatientHistory/PatientCard/PatientCard.vue
 import TreatmentCard from '@/components/PatientHistory/TreatmentCard/TreatmentCard.vue';
 import { usePatientCardStore } from '@/stores/patientCardStore';
 import { useTreatmentStore } from '@/stores/treatmentStore';
+import { useMedicalHistoryStore } from '@/stores/medicalHistoryStore';
 
 export default {
     name: 'PatientHistory',
@@ -20,7 +19,6 @@ export default {
         return {
             loading: true,
             error: null,
-            medicalData: [],
             showTestModal: false,
             showAppointmentInlineModal: false
         };
@@ -31,6 +29,9 @@ export default {
         },
         treatmentStore() {
             return useTreatmentStore();
+        },
+        medicalData() {
+            return useMedicalHistoryStore().events;
         }
     },
     watch: {
@@ -52,7 +53,6 @@ export default {
             this.loading = true;
             useAppointmentAddStore().setTreatmentActive(false);
             this.error = null;
-            this.medicalData = [];
 
             try {
                 const treatmentStore = useTreatmentStore();
@@ -65,78 +65,9 @@ export default {
                 const isActive = treatment.realEndDt === null || treatment.realEndDt === undefined;
                 useAppointmentAddStore().setTreatmentActive(isActive);
 
-                let historyItems = [];
-                let appointments = [];
                 if (treatment['@id']) {
-                    const apptResp = await apiClient.get('/appointments', {
-                        params: {
-                            treatment: treatment['@id'],
-                            order: { appointmentDt: 'asc' },
-                            itemsPerPage: 1000
-                        }
-                    });
-                    appointments = apptResp.data.member || [];
-
-                    const historyResp = await testHistoryApi.getAll({
-                        treatment: treatment['@id'],
-                        order: { creationDt: 'desc' },
-                        itemsPerPage: 300
-                    });
-                    historyItems = historyResp.member || [];
+                    await useMedicalHistoryStore().fetchMedicalData(treatment['@id']);
                 }
-
-                const events = [];
-
-                historyItems.forEach(item => {
-                    const testDate = new Date(item.creationDt);
-                    const testDateStr = testDate.toLocaleDateString('ru-RU');
-                    const matchingAppt = appointments.find(a => {
-                        const apptDate = new Date(a.appointmentDt);
-                        return apptDate.toLocaleDateString('ru-RU') === testDateStr;
-                    });
-                    if (matchingAppt) {
-                        appointments = appointments.filter(a => a !== matchingAppt);
-                    }
-                    events.push({
-                        type: 'test',
-                        date: item.creationDt,
-                        inr: item.mno !== undefined ? item.mno : '—',
-                        currentDose: item.doze !== undefined ? item.doze : '—',
-                        prescribedDose: matchingAppt ? matchingAppt.doze : '—',
-                        recommendations: matchingAppt ? (matchingAppt.comment || '') : '',
-                        comment: item.comment || ''
-                    });
-                });
-
-                appointments.forEach(a => {
-                    events.push({
-                        type: 'appointment',
-                        date: a.appointmentDt,
-                        inr: '—',
-                        currentDose: '—',
-                        prescribedDose: a.doze,
-                        recommendations: a.comment || '',
-                        comment: ''
-                    });
-                });
-
-                events.sort((a, b) => new Date(b.date) - new Date(a.date));
-
-                this.medicalData = events.map(event => ({
-                    date: event.date
-                        ? new Date(event.date).toLocaleDateString('ru-RU')
-                        : '—',
-                    inr: event.inr,
-                    analysis1: '—',
-                    analysis2: '—',
-                    heartRate: '—',
-                    bloodPressure: '—',
-                    currentDose: event.currentDose,
-                    prescribedDose: event.prescribedDose,
-                    recommendations: event.recommendations,
-                    comment: event.comment,
-                    isAppointment: event.type === 'appointment'
-                }));
             } catch (err) {
                 console.error('Ошибка загрузки истории:', err);
                 this.error = 'Не удалось загрузить данные пациента.';
